프리뷰 화면 api 리스폰스 연결

+++ app/src/main/java/com/mommydndn/app/ui/components/box/ReviewBox.kt
@@ -33,7 +33,7 @@ import com.mommydndn.app.ui.theme.caption200
 @Composable
 fun ReviewBox(
     modifier: Modifier = Modifier,
-    dndnScore: Double = 0.0,
+    dndnScore: Int = 0,
     titleText: String = "",
     dateText: String = "",
     contentText: String = "",
